+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Please choose versions by [Semantic Versioning](http://semver.org/).
 * MINOR version when you add functionality in a backwards-compatible manner, and
 * PATCH version when you make backwards-compatible bug fixes.
 
+## v1.12.0
+
+- add ConsumerOptions to OffsetConsumer
+- add ConsumerOptions to SimpleConsumer
+- allow wait if lag is small
+
 ## v1.11.3
 
 - fix NewMessageHandlerBatchDelay